Rename function `icon` to `render_icon`
