+++ src/django_icons/__init__.py
@@ -1,41 +1,4 @@
-from django.utils.functional import keep_lazy_text
-
 from .__about__ import __version__
-from .utils import get_icon_kwargs, get_icon_renderer
+from .core import icon, render_icon
 
-__all__ = ["__version__", "icon"]
-
-
-@keep_lazy_text
-def icon(name, *args, **kwargs):
-    """
-    Render an icon.
-
-    **Parameters**:
-
-        name
-            The name of the icon to be rendered
-
-        title
-            The title attribute for the icon
-
-            :default: None (no title attribute rendered)
-
-        renderer
-            The renderer to use for the icon
-
-            :default: The default renderer as per ``settings.py``, or ultimately `FontAwesome4Renderer`.
-
-    **Usage**::
-
-        icon(name)
-
-    **Example**::
-
-        icon('pencil')
-        icon('trash', title='Delete')
-    """
-    icon_kwargs = get_icon_kwargs(name, *args, **kwargs)
-    renderer_class = get_icon_renderer(icon_kwargs.get("renderer", None))
-    renderer = renderer_class(**icon_kwargs)
-    return renderer.render()
+__all__ = ["__version__", "icon", "render_icon"]
