+++ tests/test_core.py
@@ -2,12 +2,12 @@ from django.test import TestCase
 from django.utils import translation
 from django.utils.translation import gettext_lazy as _
 
-from django_icons import icon
+from django_icons.core import icon, render_icon
 from django_icons.renderers import IconRenderer
 
 
 class VersionTest(TestCase):
-    """Test the IconRenderer."""
+    """Test the package version."""
 
     def test_version(self):
         from django_icons import __version__
@@ -16,14 +16,27 @@ class VersionTest(TestCase):
         self.assertEqual(len(parts), 2)
 
 
-class IconFunctionTest(TestCase):
-    """Test the icon function."""
+class RenderIconFunctionTest(TestCase):
+    """Test the `render_icon` function."""
 
-    def test_laziness(self):
+    def test_render_icon(self):
         self.assertEqual(
-            icon("user", title=_("user"), renderer=IconRenderer),
+            render_icon("user", title=_("user"), renderer=IconRenderer),
             '<i class="user" title="user"></i>',
         )
-        user_icon = icon("user", title=_("user"), renderer=IconRenderer)
+
+    def test_laziness(self):
+        user_icon = render_icon("user", title=_("user"), renderer=IconRenderer)
         with translation.override("nl"):
             self.assertEqual(user_icon, '<i class="user" title="gebruiker"></i>')
+
+
+class IconFunctionTest(TestCase):
+    """Test the `icon` function."""
+
+    def test_icon(self):
+        with self.assertWarns(Warning):
+            self.assertEqual(
+                icon("user", title=_("user"), renderer=IconRenderer),
+                '<i class="user" title="user"></i>',
+            )
